# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * BolForm API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProviderStatusChat = typeof ProviderStatusChat[keyof typeof ProviderStatusChat];


export const ProviderStatusChat = {
  ready: 'ready',
  blocked: 'blocked',
  unchecked: 'unchecked',
} as const;

export type ProviderStatusSpeech = typeof ProviderStatusSpeech[keyof typeof ProviderStatusSpeech];


export const ProviderStatusSpeech = {
  ready: 'ready',
  blocked: 'blocked',
  unchecked: 'unchecked',
} as const;

export type ProviderStatusTranscription = typeof ProviderStatusTranscription[keyof typeof ProviderStatusTranscription];


export const ProviderStatusTranscription = {
  ready: 'ready',
  blocked: 'blocked',
  unchecked: 'unchecked',
} as const;

export type ProviderStatusDocumentAi = typeof ProviderStatusDocumentAi[keyof typeof ProviderStatusDocumentAi];


export const ProviderStatusDocumentAi = {
  ready: 'ready',
  blocked: 'blocked',
  unchecked: 'unchecked',
} as const;

export interface ProviderStatus {
  configured: boolean;
  chat: ProviderStatusChat;
  speech: ProviderStatusSpeech;
  transcription: ProviderStatusTranscription;
  documentAi: ProviderStatusDocumentAi;
  message: string;
}

export type FormFieldType = typeof FormFieldType[keyof typeof FormFieldType];


export const FormFieldType = {
  short_text: 'short_text',
  multiline: 'multiline',
  date: 'date',
  number: 'number',
  email: 'email',
  phone: 'phone',
  select: 'select',
  checkbox: 'checkbox',
} as const;

export type FormFieldRequired = typeof FormFieldRequired[keyof typeof FormFieldRequired];


export const FormFieldRequired = {
  required: 'required',
  optional: 'optional',
  unknown: 'unknown',
} as const;

export interface FormField {
  id: string;
  label: string;
  type: FormFieldType;
  options: string[];
  required: FormFieldRequired;
  instructions: string;
  constraint: string;
}

export interface FormSchema {
  id: string;
  title: string;
  sourceLanguage: string;
  instructions: string;
  fictional: boolean;
  /** @maxItems 25 */
  fields: FormField[];
}

export interface ParseTextInput {
  /**
     * @minLength 3
     * @maxLength 12000
     */
  text: string;
}

export type FieldValueStatus = typeof FieldValueStatus[keyof typeof FieldValueStatus];


export const FieldValueStatus = {
  answered: 'answered',
  needs_confirmation: 'needs_confirmation',
  skipped: 'skipped',
} as const;

export interface FieldValue {
  fieldId: string;
  value: unknown;
  status: FieldValueStatus;
}

export type ConversationInputLanguage = typeof ConversationInputLanguage[keyof typeof ConversationInputLanguage];


export const ConversationInputLanguage = {
  'hi-IN': 'hi-IN',
  'en-IN': 'en-IN',
} as const;

export interface ConversationInput {
  turnId: string;
  /** @minimum 0 */
  revision: number;
  language: ConversationInputLanguage;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  utterance: string;
  /** @maxLength 1000 */
  currentQuestion: string;
  schema: FormSchema;
  values: FieldValue[];
}

export type FieldPatchStatus = typeof FieldPatchStatus[keyof typeof FieldPatchStatus];


export const FieldPatchStatus = {
  answered: 'answered',
  needs_confirmation: 'needs_confirmation',
  skipped: 'skipped',
} as const;

export interface FieldPatch {
  fieldId: string;
  value: unknown;
  status: FieldPatchStatus;
  evidence: string;
}

export interface ConversationResult {
  turnId: string;
  revision: number;
  reply: string;
  nextQuestion: string;
  patches: FieldPatch[];
  complete: boolean;
}

export type SpeechInputLanguage = typeof SpeechInputLanguage[keyof typeof SpeechInputLanguage];


export const SpeechInputLanguage = {
  'hi-IN': 'hi-IN',
  'en-IN': 'en-IN',
} as const;

export interface SpeechInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  text: string;
  language: SpeechInputLanguage;
}

export interface SpeechResult {
  audioBase64: string;
  mimeType: string;
}

export type StreamSpeechParams = {
/**
 * @minLength 1
 * @maxLength 500
 */
text: string;
language: StreamSpeechLanguage;
};

export type StreamSpeechLanguage = typeof StreamSpeechLanguage[keyof typeof StreamSpeechLanguage];


export const StreamSpeechLanguage = {
  'hi-IN': 'hi-IN',
  'en-IN': 'en-IN',
} as const;